# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/notebook/": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Получение всех блокнотов",
                "description": " На одной странице отображается 5 блокнотов",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Добавление блокнота",
                "description": "Store notebook in DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "third_name",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Имя"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Фамилия"
                                    },
                                    "third_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Отчество"
                                    },
                                    "company": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Компания"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "ххххххххххх"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Почта"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "Y-m-d",
                                        "example": "хххх-хх-хх"
                                    },
                                    "photo": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Получение блокнота по id",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Редактирование блокнота по id",
                "description": "Update notebook in DB",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Имя"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Фамилия"
                                    },
                                    "third_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Отчество"
                                    },
                                    "company": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Компания"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "ххххххххххх"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Почта"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "Y-m-d",
                                        "example": "хххх-хх-хх"
                                    },
                                    "photo": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Удаление блокнота по id",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}